Task #4: Support parsing int64_t and uint32_t values directly
## What changed
parse_integer previously always used int32_t as the intermediate parsing
type, meaning values > 2^31-1 were silently truncated before clamp_to_max
could act on them.

### src/JSONStreamParser.h
- parse_integer now dispatches based on the target type V:
  - sizeof(V) > 4 && unsigned  → parse_numeric<V, uint64_t>
  - sizeof(V) > 4 && signed    → parse_numeric<V, int64_t>
  - sizeof(V) == 4 && unsigned → parse_numeric<V, uint64_t>  (uint32_t case)
  - everything else            → parse_numeric<V, int32_t>   (original path)
  The std::is_integral_v<V> guard prevents non-numeric types such as
  JSONCallbackObject (which is sizeof > 4 but not arithmetic) from being
  incorrectly routed to the 64-bit path.
- parse_numeric gained two new constexpr branches:
  - int64_t  → std::strtoll
  - uint64_t → std::strtoull
- The trailing-dot guard was generalised from is_same_v<Type,int32_t> to
  is_integral_v<Type> so it covers all integer intermediates.

### examples/JSONParserTest/test.h
- BigStruct f05 (int64_t) JSON value updated: 999999 → 4000000000
- BigStruct f08 (uint32_t) JSON value updated: 300000 → 3000000000
- Corresponding check() assertions updated to 4000000000LL / 3000000000U

## Result
All 131 tests pass (Release build, Clang 17).
No new compiler warnings introduced.


## No deviations from task spec

+++ src/utils.h
@@ -36,7 +36,7 @@ template <typename T> bool get_unsigned_integral_fromHexString(T &value, const c
 
 template <typename T> constexpr T be_to_h(T value);
 
-uint64_t now();
+unsigned long long now();
 
 constexpr uint8_t _hex_to_dec(char c) {
   if (c >= '0' && c <= '9') {
@@ -167,7 +167,7 @@ void print_bitwise_mask(size_t mask, size_t count) {
   DEBUG_PRINTF("\n");
 }
 
-uint64_t now() {
+unsigned long long now() {
   auto now = std::chrono::steady_clock::now();
   return std::chrono::duration_cast<std::chrono::microseconds>(now.time_since_epoch()).count();
 }
